requirements.txt
chardet
python-magic
